lateshow-firstname-lastname/requirements.txt
Flask==2.3.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
pytest==8.2.1